--- scripts/app/configs/webpack.config.development.ts
@@ -1,3 +1,0 @@
-import clientConfig from "./webpack.config.client.development";
-
-export default [clientConfig];
